Add in init and set up files. Rename files from toy to lp
