+++ penaltymodel_lp/penaltymodel/lp/__init__.py
@@ -1,0 +1,5 @@
+from penaltymodel.mip.generation import *
+import penaltymodel.mip.generation
+
+from penaltymodel.lp.interface import *
+import penaltymodel.lp.interface
